chore: 🤖 separate main file and modify Makefile
