+++ parser.c
@@ -1,0 +1,129 @@
+#include "main.h"
+
+// -------------------
+// Parser
+// -------------------
+
+// AST node type
+typedef struct Node Node;
+
+// Create New Node
+// return created new node
+Node *new_node(enum Nodekind kind, Node *left, Node *right) {
+    Node *node = calloc(1, sizeof(Node));
+    node->kind = kind;
+    node->right = right;
+    node->left = left;
+    return node;
+}
+
+// Create New node has integer
+// return created new node
+Node *new_node_num(int val){
+    Node *node = calloc(1, sizeof(Node));
+    node->kind = ND_NUM;
+    node->val = val;
+    return node;
+}
+
+// prototype declare
+Node *expr();
+Node *equality();
+Node *relational();
+Node *add();
+Node *mul();
+Node *unary();
+Node *primary();
+
+// expr = mul ("+" mul | "-" mul)*
+// The Seventh priority
+Node *expr() {
+    return equality();
+}
+
+// equality   = relational ( "==" relational | "!=" relational )*
+// The Sixth priority
+Node *equality() {
+    Node *node = relational();
+
+    for(;;) {
+        if (consume("=="))
+            node = new_node(ND_EQ, node, relational());
+        else if (consume("!="))
+           node = new_node(ND_NE, node, relational());
+        else
+            return node;
+    }
+}
+
+// relational = add ( "<" add | "<=" add | ">" add | ">=" add)*
+// The Fifth priority
+Node *relational() {
+    Node *node = add();
+
+    for(;;) {
+        if (consume("<"))
+            node = new_node(ND_LT, node, add());
+        else if (consume("<="))
+            node = new_node(ND_LE, node, add());
+        else if (consume(">"))
+            node = new_node(ND_LT, add(), node);
+        else if (consume(">="))
+            node = new_node(ND_LE, add(), node);
+        else
+            return node;
+    }
+}
+
+// add = mul ("+" mul | "-" mul)*
+// The Fouth priority
+Node *add() {
+    Node *node = mul();
+
+    for(;;) {
+        if (consume("+"))
+            node = new_node(ND_ADD, node, mul());
+        else if (consume("-"))
+            node = new_node(ND_SUB, node, mul());
+        else
+            return node;
+    }
+}
+
+// mul = unary ("*" unary | "/" unary)*
+// The Third priority
+Node *mul() {
+    Node *node = unary();
+
+    for(;;) {
+        if (consume("*"))
+            node = new_node(ND_MUL, node, unary());
+        else if (consume("/"))
+            node = new_node(ND_DIV, node, unary());
+        else
+            return node;
+    }
+}
+
+// unary = ("+" | "-")? unary
+//         | primary
+// The Second priority
+Node *unary() {
+    if (consume("+"))
+        return unary();
+    if (consume("-"))
+        return new_node(ND_SUB, new_node_num(0), unary());
+    return primary();
+}
+
+// primary = num | "(" expr ")"
+// The First Priority
+Node *primary() {
+    if (consume("(")) {
+        Node *node = expr();
+        expect(")");
+        return node;
+    }
+
+    return new_node_num(expect_number());
+}
